python_service/requirements.txt
fastapi
uvicorn
face_recognition
numpy
pillow
pydantic
requests
python-multipart
